+++ integration/lang_test.go
@@ -933,3 +933,4 @@ func TestExecImportReversedString(t *testing.T) {
 		_ = os.Getwd
 }
 
+